refactor
- adjust module structure
- remove unused method with `todo!()`
- no recursion in `next()`
- baseline comparisons
