+++ gix-traverse/tests/traverse.rs
@@ -5,5 +5,4 @@ fn hex_to_id(hex: &str) -> gix_hash::ObjectId {
 }
 
 mod commit;
-mod topo;
 mod tree;
